+++ tests/conftest.py
@@ -21,12 +21,16 @@ class MockBroker(BaseBroker):
     def __init__(self, app: "Pynenc") -> None:
         super().__init__(app)
         self.route_invocation_mock = MagicMock()
+        self.route_invocations_mock = MagicMock()
         self.retrieve_invocation_mock = MagicMock()
         self.count_invocations_mock = MagicMock()
         self.purge_mock = MagicMock()
 
     def route_invocation(self, *args: Any, **kwargs: Any) -> None:
         self.route_invocation_mock(*args, **kwargs)
+
+    def route_invocations(self, *args: Any, **kwargs: Any) -> None:
+        self.route_invocations_mock(*args, **kwargs)
 
     def retrieve_invocation(self, *args: Any, **kwargs: Any) -> Any:
         return self.retrieve_invocation_mock(*args, **kwargs)
@@ -60,6 +64,7 @@ class MockBaseOrchestrator(BaseOrchestrator):
         self._get_existing_invocations_mock = MagicMock()
         self._get_invocation_mock = MagicMock()
         self._set_invocation_status_mock = MagicMock()
+        self._set_invocations_status_mock = MagicMock()
         self._set_invocation_pending_status_mock = MagicMock()
         self._get_invocation_status_mock = MagicMock()
         self._increment_invocation_retries_mock = MagicMock()
@@ -78,6 +83,9 @@ class MockBaseOrchestrator(BaseOrchestrator):
 
     def _set_invocation_status(self, *args: Any, **kwargs: Any) -> Any:
         return self._set_invocation_status_mock(*args, **kwargs)
+
+    def _set_invocations_status(self, *args: Any, **kwargs: Any) -> Any:
+        return self._set_invocations_status_mock(*args, **kwargs)
 
     def _set_invocation_pending_status(self, *args: Any, **kwargs: Any) -> Any:
         return self._set_invocation_pending_status_mock(*args, **kwargs)